SE AGREGO EL MODULO DE APARTAMENTO, PABELLON, ESTACIONAMIENTO

+++ resources/views/livewire/copropietario/copropietario-livewire.blade.php
@@ -4,7 +4,7 @@
             <div class="card-header">
                 <div class="row d-flex justify-content-between">
                     <div class="col-md-1">
-                        @can('agregar-funcionario')
+                        @can('agregar-copropietario')
                             <button class="btn btn-success btn-md" wire:click="$toggle('openModalNew')"><i class="fas fa-plus-square"></i></button>
                         @endcan
                     </div>
@@ -29,9 +29,9 @@
                                 <th>APARTAMENTO</th>
                                 <th>GENERO</th>
                                 <th>PAIS</th>
-                                {{-- @can('funciones-funcionario') --}}
+                                @can('funciones-copropietario')
                                     <th>ACCIONES</th>
-                                {{-- @endcan --}}
+                                @endcan
                             </tr>
                         </thead>
                         <tbody>
@@ -42,20 +42,20 @@
                                     <td>{{$copropietario->numero_apartamento}}</td>
                                     <td class="text-center">{{$copropietario->genero}}</td>
                                     <td class="text-center">{{$copropietario->pais}}</td>
-                                    {{-- @can('funciones-funcionario') --}}
+                                    @can('funciones-copropietario')
                                         <td class="text-center">
-                                            {{-- @can('editar-funcionario') --}}
+                                            @can('editar-copropietario')
                                                 <button class="btn btn-warning btn-sm" wire:click="edit({{ $copropietario->id }})"><i class="fas fa-pencil-alt"></i></button>
-                                            {{-- @endcan --}}
-                                            {{-- @can('eliminar-funcionario') --}}
+                                            @endcan
+                                            @can('eliminar-copropietario')
                                                 @if ($copropietario->estado == 1)
                                                     <button class="btn btn-danger btn-sm" wire:click="$dispatch('confirmDelete', {{ $copropietario->id }})"><i class="fas fa-trash"></i></button>
                                                 @else
                                                     <button class="btn btn-primary btn-sm" wire:click="$dispatch('confirmDelete', {{ $copropietario->id }})"><i class="fas fa-history"></i></button>
                                                 @endif
-                                            {{-- @endcan --}}
+                                            @endcan
                                         </td>
-                                    {{-- @endcan --}}
+                                    @endcan
                                 </tr>
                             @endforeach
                         </tbody>
@@ -72,7 +72,7 @@
         </div>
 
 
-        {{-- * MODAL --}}
+    {{-- * MODAL --}}
     @if ($openModalNew)
     <!-- Modal -->
     <div class="modal bd-example-modal-lg" tabindex="-1" aria-labelledby="exampleModalLabel" style="display: block" aria-modal="true" role="dialog">
@@ -134,7 +134,11 @@
                                 <select wire:model="copropietario.id_apartamento" class="form-control form-control-sm" required>
                                     <option value="" selected>Seleccionar</option>
                                     @foreach ($apartamentos as $apartamento)
-                                        <option value='{{$apartamento->id}}'>{{$apartamento->numero_apartamento}}</option>
+                                        @if($apartamento->estado == 1)
+                                            <option value='{{$apartamento->id}}'>{{$apartamento->numero_apartamento}}</option>
+                                        @else
+                                            <option class="text-danger" value='{{$apartamento->id}}' disabled>{{$apartamento->numero_apartamento}} - OCUPADO</option>
+                                        @endif
                                     @endforeach
                                 </select>
                             </div>
@@ -171,4 +175,109 @@
         </div>
     </div>
 @endif
+
+ {{-- * MODAL --}}
+ @if ($openModalEdit)
+ <!-- Modal -->
+ <div class="modal bd-example-modal-lg" tabindex="-1" aria-labelledby="exampleModalLabel" style="display: block" aria-modal="true" role="dialog">
+     <div class="modal-dialog modal-lg">
+         <div class="modal-content">
+
+             <!-- Encabezado del Modal -->
+             <div class="modal-header bg-primary">
+                 <h4 class="modal-title font-italic font-weight-bold">EDITAR COMPROPIETARIO</h4>
+                 <button type="button" class="btn btn-danger btn-sm buttonCerrarModal" data-dismiss="modal" wire:click="closeModal">×</button>
+             </div>
+ 
+             <!-- Contenido del Modal -->
+             <div class="modal-body">
+                 <form wire:submit="update">
+                     <div class="row">
+                         {{-- <div class="col-md-12">
+                             <input type="hidden" wire:model="obtenerIdPersona" required>
+                             <label for="">Buscar Persona*:</label>
+                             <input type="text" class="form-control buscar" placeholder="Buscar..." aria-label="Buscador" wire:model.live="searchPersona" 
+                             @if($selectedPersona) disabled @endif>
+                         </div> --}}
+                         {{-- <div class="col-md-12 rounded p-3 mt-2" id="contenedorBuscador">
+                             @if(isset($personas) && strlen($searchPersona) > 0)
+                                 @if($personas->isNotEmpty())
+                                     <table class="table table-bordered table-sm table-hover">
+                                         <thead class="bg-secondary">
+                                             <tr class="text-center">
+                                                 <th>NOMBRE</th>
+                                                 <th>CI</th>
+                                                 <th>ACCION</th>
+                                             </tr>
+                                         </thead>
+                                         <tbody>
+                                             @foreach($personas as $persona)
+                                                 <tr class="persona-row" wire:key="persona-{{$persona->id}}" data-id="{{$persona->id}}" @if($selectedPersona && $obtenerIdPersona != $persona->id) style="display: none;" @endif>
+                                                     <td>{{ $persona->nombre }} {{ $persona->apellido }}</td>
+                                                     <td class="text-center">{{ $persona->ci }}</td>
+                                                     <td class="text-center">
+                                                         <a wire:click="seleccionarPersona({{ $persona->id }})" class="btn @if($obtenerIdPersona == $persona->id) btn-danger @else btn-outline-success @endif btn-sm botones ">
+                                                             @if($obtenerIdPersona == $persona->id)
+                                                                 Cancelar
+                                                             @else
+                                                                 Seleccionar
+                                                             @endif
+                                                         </a>
+                                                     </td>
+                                                 </tr>
+                                             @endforeach
+                                         </tbody>
+                                     </table>
+                                 @else
+                                     <p class="text-center">Ningun dato encontrado.</p>
+                                 @endif
+                             @endif
+                         </div> --}}
+                         <div class="col-md-12">
+                             <label for="">Apartamento*:</label>
+                             <select wire:model="copropietario.id_apartamento" class="form-control form-control-sm" required>
+                                 <option value="" selected>Seleccionar</option>
+                                 @foreach ($apartamentos as $apartamento)
+                                    @if($apartamento->estado == 1)
+                                        <option value='{{$apartamento->id}}'>{{$apartamento->numero_apartamento}}</option>
+                                    @else
+                                        <option class="text-danger" value='{{$apartamento->id}}'>{{$apartamento->numero_apartamento}} - OCUPADO</option>
+                                    @endif
+                                 @endforeach
+                             </select>
+                         </div>
+                         <div class="col-md-6 mt-4">
+                             <label for="">CANTIDAD RESIDENTE*:</label>
+                             <input type="number" wire:model="copropietario.cant_residentes" class="form-control form-control-sm" required>
+                         </div>
+                         <div class="col-md-6 mt-4">
+                             <label for="">CANTIDAD MASCOTA:</label>
+                             <input type="number" wire:model="copropietario.cant_mascotas" class="form-control form-control-sm" value="0" required>
+                         </div>
+                         <div class="col-md-12 mt-4 validate">
+                             @if ($errors->any())
+                             <div class="alert alert-danger">
+                                     <ul>
+                                         @foreach ($errors->all() as $error)
+                                             <li>{{ $error }}</li>
+                                         @endforeach
+                                     </ul>
+                                 </div>
+                             @endif
+                         </div>
+                     </div>
+           
+                     <div class="row d-flex justify-content-end mt-4">
+                         <x-button class="btn-success btn-sm">Guardar</x-button>
+                     </div>
+                 </form>
+             </div>
+  
+              
+              
+         </div>
+     </div>
+ </div>
+@endif
+
 </div>
